pyproject.toml
[tool.poetry]
name = "vaton"
version = "0.1.0"
description = ""
authors = ["iz_work <iz@monosnap.com>"]

[tool.poetry.dependencies]
python = "^3.9"
isort = "^5.10.1"
autoflake = "^1.4"
black = "^22.3.0"
mypy = "^0.942"
flake8 = "^4.0.1"
bandit = "^1.7.4"
safety = "^1.10.3"
uvicorn = "^0.17.6"
pytest = "^7.1.1"
SQLAlchemy = "^1.4.35"
asyncpg = "^0.25.0"
alembic = "^1.7.7"
PyJWT = "^2.3.0"
types-toml = "^0.10.4"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.75.2"
zxcvbn = "^4.4.28"
cryptocode = "^0.1"
odmantic = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.9"
exclude = [
    '^app/alembic/*',
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "sqlalchemy.*",
    "base_class",
    "sqlalchemy",
    "jose",
    "motor.motor_asyncio",
    'passlib.context',
    'cryptocode',
    'zxcvbn'
]


